refactor: headers 관련 fetch 함수로 분리
