--- src/apis/auth.ts
@@ -1,6 +1,0 @@
-const email = process.env.REACT_APP_EMAIL;
-const password = process.env.REACT_APP_PASSWORD;
-
-const credentials = btoa(email + ':' + password);
-
-export default credentials;
