+++ src/apis/products.ts
@@ -2,15 +2,13 @@ import type { ProductType } from '../types/product';
 import type { HostNameType } from '../types/server';
 
 import { servers } from '../constants/server';
+import fetchWithHeaders from '.';
 
 export const getProducts = async (hostName: HostNameType) => {
   const hostURL = servers[hostName];
-  const response = await fetch(`${hostURL}/products`);
 
-  if (!response.ok) {
-    throw new Error(response.status.toString());
-  }
-
+  const response = await fetchWithHeaders(`${hostURL}/products`, 'GET');
   const data: ProductType[] = await response.json();
+
   return data;
 };
